Change ByteCode to byte
made little sense to be otherwise.

+++ src/edu/mccc/cos210/ds/fp/javaforth/machineModel/Instruction.java
@@ -27,7 +27,7 @@ public enum Instruction {
 	Instruction(int hexValue){
 		this.byteCode = hexValue;
 	}
-	public int getByteCode() {
-		return byteCode;
+	public Byte getByteCode() {
+		return (byte) byteCode;
 	}
 }
